refactor(frontend): decompose HomeDashboardParts.tsx God File
Extracted sub-components from HomeDashboardParts.tsx into individual files inside `frontend/src/components/home/dashboard/` to resolve the Mega Widget violation. Updated `home.tsx` to pull all these components from the barrel file `@/components/home/index.ts`. All behavioral logic and typing is preserved perfectly.
